Extract concern form time range dependant Healthchecks
